requirements.txt
telethon==1.34.0
python-dotenv==1.0.1
pydantic==2.7.4
binance-sdk==2.2.2
ujson==5.10.0
python-telegram-bot==21.6
